Revert "finished dynamic glowshroom (no worldgen yet)"
This reverts commit 64cb2f8f
