+++ src/main/java/maxhyper/dtquark/DynamicTreesQuark.java
@@ -6,11 +6,9 @@ import com.ferreusveritas.dynamictrees.block.leaves.LeavesProperties;
 import com.ferreusveritas.dynamictrees.block.rooty.SoilProperties;
 import com.ferreusveritas.dynamictrees.tree.family.Family;
 import com.ferreusveritas.dynamictrees.tree.species.Species;
-import maxhyper.dtquark.dynamictreesplus.PlusRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.IEventBus;
-import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
@@ -31,10 +29,6 @@ public class DynamicTreesQuark {
         modEventBus.addListener(this::clientSetup);
         modEventBus.addListener(this::commonSetup);
         modEventBus.addListener(this::gatherData);
-
-        if (ModList.get().isLoaded("dynamictreesplus")){
-            modEventBus.register(PlusRegistries.class);
-        }
 
         MinecraftForge.EVENT_BUS.register(this);
