+++ src/main/resources/assets/dtquark/models/block/saplings/glow_shroom.json
