--- src/main/resources/data/dtquark/loot_tables/blocks/glow_shroom_cap.json
@@ -1,59 +1,0 @@
-{
-  "type": "minecraft:block",
-  "pools": [
-    {
-      "rolls": 1.0,
-      "bonus_rolls": 0.0,
-      "entries": [
-        {
-          "type": "minecraft:alternatives",
-          "children": [
-            {
-              "type": "minecraft:item",
-              "conditions": [
-                {
-                  "condition": "minecraft:match_tool",
-                  "predicate": {
-                    "enchantments": [
-                      {
-                        "enchantment": "minecraft:silk_touch",
-                        "levels": {
-                          "min": 1
-                        }
-                      }
-                    ]
-                  }
-                }
-              ],
-              "name": "quark_glow_shroom_block"
-            },
-            {
-              "type": "minecraft:item",
-              "functions": [
-                {
-                  "function": "minecraft:set_count",
-                  "count": {
-                    "type": "minecraft:uniform",
-                    "min": -6.0,
-                    "max": 2.0
-                  },
-                  "add": false
-                },
-                {
-                  "function": "minecraft:limit_count",
-                  "limit": {
-                    "min": 0.0
-                  }
-                },
-                {
-                  "function": "minecraft:explosion_decay"
-                }
-              ],
-              "name": "quark:glow_shroom"
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}
