--- src/main/resources/data/dtquark/loot_tables/trees/mushroom_caps/glow_shroom.json
@@ -1,55 +1,0 @@
-{
-  "pools": [
-    {
-      "rolls": 1.0,
-      "bonus_rolls": 0.0,
-      "entries": [
-        {
-          "type": "minecraft:alternatives",
-          "children": [
-            {
-              "type": "minecraft:item",
-              "conditions": [
-                {
-                  "condition": "minecraft:match_tool",
-                  "predicate": {
-                    "enchantments": [
-                      {
-                        "enchantment": "minecraft:silk_touch",
-                        "levels": {
-                          "min": 1
-                        }
-                      }
-                    ]
-                  }
-                }
-              ],
-              "name": "quark:glow_shroom_block"
-            },
-            {
-              "type": "minecraft:item",
-              "name": "quark:glow_shroom"
-            }
-          ]
-        }
-      ],
-      "functions": [
-        {
-          "function": "minecraft:explosion_decay"
-        }
-      ],
-      "conditions": [
-        {
-          "condition": "minecraft:table_bonus",
-          "enchantment": "minecraft:fortune",
-          "chances": [
-            0.2,
-            0.2333333,
-            0.2666666,
-            0.3
-          ]
-        }
-      ]
-    }
-  ]
-}
