+++ src/main/resources/trees/dtquark/world_gen/feature_cancellers.json
@@ -1,11 +1,15 @@
 [
   {
-    "_comment": "Disabled for now, underground generation is WIP",
-    "_only_if_loaded": "dynamictreesplus",
-    "_select": { "name": "quark:glimmering_weald" },
-    "_cancellers": {
+    "only_if_loaded": "dynamictreesplus",
+    "select": {
+      "name": "quark:glimmering_weald"
+    },
+    "cancellers": {
       "type": "dtquark:mushroom",
-      "namespaces": ["minecraft", "quark"],
+      "namespaces": [
+        "minecraft",
+        "quark"
+      ],
       "stage": "UNDERGROUND_DECORATION"
     }
   }
